non-linear-correlation-analysis/requirements.txt
pandas>=2.0
numpy>=1.24,<2.0
scipy>=1.11,<1.12
scikit-learn>=1.3,<1.4
matplotlib>=3.7,<3.8
seaborn>=0.13,<0.14
openpyxl>=3.1
PyYAML>=6.0
